+++ src/js/squadSimulation.js
@@ -1,6 +1,6 @@
 import { App } from "../app.js";
-import targetIcon from "../img/icons/marker_target_enabled.webp";
-import targetIconDisabled from "../img/icons/marker_target_disabled.webp";
+import targetIcon from "../img/icons/marker_target_enabled_xmas.webp";
+import targetIconDisabled from "../img/icons/marker_target_disabled_xmas.webp";
 import i18next from "i18next";
 
 export default class Simulation {